Move errors + remove exports
