+++ src/4.3_to_5.0/errors/UnsupportedLegacyChartTypeError.ts
