+++ src/4.3_to_5.0/index.ts
@@ -1,9 +1,1 @@
-export { migrateDashboard } from "./migrateDashboard";
-export { migrateWidget } from "./migrateWidget";
-export { migrateChart } from "./migrateChart";
-export { migrateTable } from "./migrateTable";
-export { migrateKpi } from "./migrateKpi";
-export { migrateQuickFilter } from "./migrateQuickFilter";
-export { migrateUIFolder43To50 } from "./migrateUIFolder43To50";
-
-export * from "./migration.types";
+export { migrate_43_to_50 } from "./migrate_43_to_50";
