+++ src/4.3_to_5.0/migrateDashboard.ts
@@ -25,8 +25,8 @@ import { isLegacyLayoutLeaf } from "./isLegacyLayoutLeaf";
 import { _migrateContextValues } from "./_migrateContextValues";
 import { _getLegacyWidgetPluginKey } from "./_getLegacyWidgetPluginKey";
 import { _serializeError } from "./_serializeError";
-import { PartialMigrationError } from "../errors/PartialMigrationError";
-import { WidgetFlaggedForRemovalError } from "../errors/WidgetFlaggedForRemovalError";
+import { PartialMigrationError } from "./errors/PartialMigrationError";
+import { WidgetFlaggedForRemovalError } from "./errors/WidgetFlaggedForRemovalError";
 
 /**
  * Adds `error` to `errorReport`, where `error` was thrown during the migration of a widget within the dashboard.
