+++ src/4.3_to_5.0/migrateDrillthrough.ts
@@ -5,7 +5,7 @@ import type {
   DrillthroughTableWidgetState,
 } from "@activeviam/activeui-sdk-5.0";
 import { serializeWidgetState, parse } from "@activeviam/activeui-sdk-5.0";
-import { UnsupportedLegacyQueryUpdateModeError } from "../errors/UnsupportedLegacyQueryUpdateModeError";
+import { UnsupportedLegacyQueryUpdateModeError } from "./errors/UnsupportedLegacyQueryUpdateModeError";
 import { _getQueryInLegacyWidgetState } from "./_getQueryInLegacyWidgetState";
 import { _getTargetCubeFromServerUrl } from "./_getTargetCubeFromServerUrl";
 import { _migrateQuery } from "./_migrateQuery";
