+++ src/4.3_to_5.0/migrateKpi.ts
@@ -15,7 +15,7 @@ import {
   deriveMappingFromMdx,
 } from "@activeviam/activeui-sdk-5.0";
 import { getSpecificCompoundIdentifier } from "@activeviam/mdx-5.0";
-import { UnsupportedLegacyQueryUpdateModeError } from "../errors/UnsupportedLegacyQueryUpdateModeError";
+import { UnsupportedLegacyQueryUpdateModeError } from "./errors/UnsupportedLegacyQueryUpdateModeError";
 import { _getQueryInLegacyWidgetState } from "./_getQueryInLegacyWidgetState";
 import { _getTargetCubeFromServerUrl } from "./_getTargetCubeFromServerUrl";
 import { _migrateQuery } from "./_migrateQuery";
