+++ src/4.3_to_5.0/migrateTable.ts
@@ -12,7 +12,7 @@ import {
   pluginWidgetTable,
   pluginWidgetTreeTable,
 } from "@activeviam/activeui-sdk-5.0";
-import { UnsupportedLegacyQueryUpdateModeError } from "../errors/UnsupportedLegacyQueryUpdateModeError";
+import { UnsupportedLegacyQueryUpdateModeError } from "./errors/UnsupportedLegacyQueryUpdateModeError";
 import { _getQueryInLegacyWidgetState } from "./_getQueryInLegacyWidgetState";
 import { _getTargetCubeFromServerUrl } from "./_getTargetCubeFromServerUrl";
 import { _migrateQuery } from "./_migrateQuery";
