+++ src/4.3_to_5.0/migrate_43_to_50.test.ts
@@ -1,6 +1,6 @@
 import _map from "lodash/map";
 import _some from "lodash/some";
-import { migrateUIFolder43To50 } from "./migrateUIFolder43To50";
+import { migrate_43_to_50 } from "./migrate_43_to_50";
 import { smallLegacyUIFolder } from "./__test_resources__/smallLegacyUIFolder";
 import { legacyUIFolder } from "./__test_resources__/legacyUIFolder";
 import { servers } from "./__test_resources__/servers";
@@ -31,31 +31,35 @@ jest.mock(`./generateId`, () => {
   };
 });
 
-describe("migrateUIFolder43To50", () => {
+describe("migrate_43_to_50", () => {
   it("returns a valid ActiveUI5 /ui folder on a small input", async () => {
-    const [migratedUIFolder, counters, errorReport] =
-      await migrateUIFolder43To50(smallLegacyUIFolder, {
+    const [migratedUIFolder, counters, errorReport] = await migrate_43_to_50(
+      smallLegacyUIFolder,
+      {
         servers,
         doesReportIncludeStacks: false,
-      });
+      },
+    );
     expect(migratedUIFolder).toMatchSnapshot();
     expect(errorReport).toBeUndefined();
     expect(counters).toMatchSnapshot();
   });
 
   it("returns a valid ActiveUI5 /ui folder on a real life input", async () => {
-    const [migratedUIFolder, counters, errorReport] =
-      await migrateUIFolder43To50(legacyUIFolder, {
+    const [migratedUIFolder, counters, errorReport] = await migrate_43_to_50(
+      legacyUIFolder,
+      {
         servers,
         doesReportIncludeStacks: false,
-      });
+      },
+    );
     expect(migratedUIFolder).toMatchSnapshot();
     expect(errorReport).toMatchSnapshot();
     expect(counters).toMatchSnapshot();
   });
 
   it("returns a valid ActiveUI5 /ui folder that includes calculated measures when the input includes a pivotFolder", async () => {
-    const [migratedUIFolder] = await migrateUIFolder43To50(legacyUIFolder, {
+    const [migratedUIFolder] = await migrate_43_to_50(legacyUIFolder, {
       servers,
       legacyPivotFolder: smallLegacyPivotFolder,
       doesReportIncludeStacks: false,
@@ -69,7 +73,7 @@ describe("migrateUIFolder43To50", () => {
 
   it("removes the specified widget plugins from the widget bookmarks themselves, and from the content of the dashboard bookmarks", async () => {
     const keysOfWidgetPluginsToRemove = ["filters"];
-    const [migratedUIFolder] = await migrateUIFolder43To50(legacyUIFolder, {
+    const [migratedUIFolder] = await migrate_43_to_50(legacyUIFolder, {
       servers,
       keysOfWidgetPluginsToRemove,
       doesReportIncludeStacks: false,
@@ -129,7 +133,7 @@ describe("migrateUIFolder43To50", () => {
   });
 
   it("returns an error report for dashboards and handles the dashboard id being a number", async () => {
-    const [migratedFolder, counters, errorReport] = await migrateUIFolder43To50(
+    const [migratedFolder, counters, errorReport] = await migrate_43_to_50(
       smallLegacyUIFolderWithInvalidDashboard,
       {
         servers,
@@ -187,7 +191,7 @@ describe("migrateUIFolder43To50", () => {
   });
 
   it("copies invalid filters as-is and reports an error", async () => {
-    const [migratedFolder, counters, errorReport] = await migrateUIFolder43To50(
+    const [migratedFolder, counters, errorReport] = await migrate_43_to_50(
       smallLegacyUIFolderWithInvalidFilter,
       {
         servers,
