+++ src/4.3_to_5.0/migrate_43_to_50.ts
@@ -25,8 +25,8 @@ import {
 import { _getFolderName } from "./_getFolderName";
 import { _getMapOfFolderIds } from "./_getMapOfFolderIds";
 import { _serializeError } from "./_serializeError";
-import { PartialMigrationError } from "../errors/PartialMigrationError";
-import { WidgetFlaggedForRemovalError } from "../errors/WidgetFlaggedForRemovalError";
+import { PartialMigrationError } from "./errors/PartialMigrationError";
+import { WidgetFlaggedForRemovalError } from "./errors/WidgetFlaggedForRemovalError";
 import cliProgress from "cli-progress";
 
 const _getFolder = (
@@ -208,7 +208,7 @@ const accumulateStructure = ({
  * - for a matching saved ActiveUI 4 widget, no ActiveUI 5 file is created.
  * - for a saved ActiveUI 4 dashboard including a matching widget, the widget is removed from the output ActiveUI 5 dashboard, and the layout is adapted so that siblings take the remaining space.
  */
-export async function migrateUIFolder43To50(
+export async function migrate_43_to_50(
   legacyUIFolder: ContentRecord,
   {
     servers,
