+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from "./4.3_to_5.0";
+export { migrate_43_to_50 } from "./4.3_to_5.0";
